Move one hot encoding function into own file
